New Arla userland from Arla between 0.34.2 and current in arla cvs. Too many new features and fixes to mention here.

+++ usr.sbin/afs/src/rx/rx_globs.h
@@ -1,5 +1,4 @@
-/*	$OpenBSD: src/usr.sbin/afs/src/rx/Attic/rx_globs.h,v 1.1.1.1 1998/09/14 21:53:15 art Exp $	*/
-/* $KTH: rx_globs.h,v 1.4 1998/02/22 19:46:16 joda Exp $ */
+/* $Id: rx_globs.h,v 1.5 1999/10/15 16:12:35 map Exp $ */
 
 /*
 ****************************************************************************
@@ -228,8 +227,8 @@ void rxi_ClearReceiveQueue(struct rx_call *);
 void rxi_ResetConnection(struct rx_connection *);
 void rxi_InitCall(void); /* obsolete ? */
 void rxi_ResetCall(struct rx_call *);
-void rxi_CallError(struct rx_call *, long);
-void rxi_ConnectionError(struct rx_connection *, long);
+void rxi_CallError(struct rx_call *, int32_t);
+void rxi_ConnectionError(struct rx_connection *, int32_t);
 void rxi_QueuePackets(void); /* obsolete ? */
 void rxi_Start(struct rxevent *, struct rx_call *);
 void rxi_CallIsIdle(void); /* obsolete ? */
